+++ NEWS.md
@@ -236,8 +236,12 @@ This section lists changes that do not have deprecation warnings.
 Library improvements
 --------------------
 
+  * Functions `first` and `last` now accept `nchar` argument for `AbstractString`.
+    If this argument is used they reutrn a substring consisting of first/last `nchar`
+    characters from the original string ([#?????]).
+
   * The functions `nextind` and `prevind` now accept `nchar` argument that indicates
-    number of characters to move ([#23805]).
+    the number of characters to move ([#23805]).
 
   * The functions `strip`, `lstrip` and `rstrip` now return `SubString` ([#22496]).